blogation add notion api

+++ cmd/blogation.go
@@ -36,10 +36,11 @@ var panicIf = log.Panic
 
 func publish(cmd *cobra.Command, args []string) {
 	fmt.Println("blogation called")
-	authToken2 := "3fae99bf3ebd3573ad80a8cc13dad050cad2094d5c03c8a6f1cde859562be8ecc392723ae668859d3a5e777d6ae4bdcb8a2fa95441f84f7bac8da9c6111e2c2ad9fbf1d168ae43392ac94f4d2e32"
+	authToken2 := "399cba2cad29b8f64f33fc0007b5145ca3486f867d6c04303feb9d1f0f857e177d8a5276f47f82cb62dc64ec7f2365538250d0b7e5b90a87061fbd7cd745c477799f5ef41919c81126c948887593"
 	pageID := "47d84abb6c0b4e0eb799eec100785fab"
-	postsDir := "/Users/admin/Documents/fencex.github.io/content/posts/"
-	imageDir := "/Users/admin/Documents/fencex.github.io/static/images"
+	workDir := os.Getenv("HANDY_WORK_DIR")
+	postsDir := workDir + "/content/"
+	imageDir := workDir + "/static/images"
 
 	client := &notionapi.Client{AuthToken: authToken2}
 	rootPage, _ := client.DownloadPage(pageID)
@@ -49,23 +50,29 @@ func publish(cmd *cobra.Command, args []string) {
 	dir, _ := ioutil.ReadDir(postsDir)
 	for _, d := range dir {
 		// todo 删除
-		os.RemoveAll(path.Join([]string{postsDir, d.Name()}...))
+		if d.IsDir() {
+			os.RemoveAll(path.Join([]string{postsDir, d.Name()}...))
+		}
 	}
 
 	for _, levelOneId := range levelOneIds {
 		// 第一级分类
-		levelOnePage, _ := client.DownloadPage(levelOneId)
+		levelOnePage, _ := client.DownloadPage(levelOneId.DashID)
 		levelOneTitle := levelOnePage.Root().GetTitle()[0].Text
+		println(levelOneTitle)
 
 		indexText := getIndexContent(levelOneTitle)
-		writeNewFile(indexText, postsDir+levelOneTitle, "_index.md")
+		if levelOneTitle != "menu" {
+			writeNewFile(indexText, postsDir+levelOneTitle, "_index.md")
+		}
 
 		levelTwoIds := levelOnePage.GetSubPages()
 		for _, levelTwoId := range levelTwoIds {
 
 			// 第二级文章
-			levelTwoPage, _ := client.DownloadPage(levelTwoId)
+			levelTwoPage, _ := client.DownloadPage(levelTwoId.DashID)
 			title := levelTwoPage.Root().GetTitle()[0].Text
+			println("\t" + title)
 
 			// 修改 markdown 文件
 			levelTwoPageMdStr := string(tomarkdown.ToMarkdown(levelTwoPage))
